Accuracy at all times... you have to put qreals into a division if you expect qreals out.
svn path=/trunk/KDE/kdebase/workspace/; revision=888162

+++ plasma/shells/plasmoidviewer/fullview.cpp
@@ -115,8 +115,8 @@ void FullView::resizeEvent(QResizeEvent *event)
     qreal newHeight = 0;
 
     if (m_applet->aspectRatioMode() == Plasma::KeepAspectRatio) {
-        qreal ratio = event->oldSize().width() / event->oldSize().height();
-        qreal newPossibleWidth = size().height() * ratio;
+        qreal ratio = (qreal)event->oldSize().width() / (qreal)event->oldSize().height();
+        qreal newPossibleWidth = (qreal)size().height() * ratio;
         if (newPossibleWidth > size().width()) {
             newHeight = size().width() / ratio;
             newWidth = newHeight * ratio;
